리액션 기능

+++ dream/src/main/java/com/example/dream/service/BoardService.java
@@ -1,15 +1,13 @@
 package com.example.dream.service;
 
 
-import com.example.dream.dto.BoardDto;
-import com.example.dream.dto.BoardListResponseDto;
-import com.example.dream.dto.BoardResponseDto;
-import com.example.dream.dto.GlobalResDto;
+import com.example.dream.dto.*;
 import com.example.dream.entity.Board;
 import com.example.dream.entity.Comment;
 import com.example.dream.entity.Member;
 import com.example.dream.repository.BoardRepository;
 import com.example.dream.repository.CommentRepository;
+import com.example.dream.repository.ReactionsRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,7 +15,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -25,15 +22,12 @@ public class BoardService {
 
     private final BoardRepository boardRepository;
      private final CommentRepository commentRepository;
-//    private final JwtUtil jwtUtil;
+
+     private final ReactionsRepository reactionsRepository;
 
 
     @Transactional
     public GlobalResDto saveBoard(BoardDto boardRequestDto, Member member) {
-
-//        if (null == member) {
-//            return new GlobalResDto("INVALID_TOKEN", HttpStatus.FORBIDDEN.value());
-//        }
 
         Board board = Board.builder()
 //                .username(boardRequestDto.getUsername())
@@ -50,11 +44,21 @@ public class BoardService {
 
     @Transactional(readOnly = true)
     public ResponseEntity<BoardResponseDto> getBoard(Long id) {
+
     // board title , memeber nickname, like count, dislike count, horror count
+
         List<Comment> commentList = commentRepository.findCommentByBoardId(id);
         Board board = checkBoard(boardRepository, id);
 
-        return ResponseEntity.ok(new BoardResponseDto(board,commentList));
+        long likeCount = reactionsRepository.countReactionsByBoard_BoardIdAndLikedTrue(id);
+        long dislikeCount = reactionsRepository.countReactionsByBoard_BoardIdAndDislikedTrue(id);
+        long horrorCount = reactionsRepository.countReactionsByBoard_BoardIdAndHorridTrue(id);
+
+        ReactionsDto dto = new ReactionsDto(likeCount,dislikeCount,horrorCount);
+
+//        BoardResponseDto boardDto = new BoardResponseDto(board, commentList,dto);
+
+        return ResponseEntity.ok(new BoardResponseDto(board,commentList,dto));
     }
 
     @Transactional(readOnly = true)
@@ -65,30 +69,43 @@ public class BoardService {
         List<Board> boards = boardRepository.findAll();
 
         for (Board board : boards) {
-            boardListResponseDto.addBoard(new BoardResponseDto(board));
-        }
+            long likeCount = reactionsRepository.countReactionsByBoard_BoardIdAndLikedTrue(board.getBoardId());
+            long dislikeCount = reactionsRepository.countReactionsByBoard_BoardIdAndDislikedTrue(board.getBoardId());
+            long horrorCount = reactionsRepository.countReactionsByBoard_BoardIdAndHorridTrue(board.getBoardId());
+            ReactionsDto dto = new ReactionsDto(likeCount,dislikeCount,horrorCount);
+            boardListResponseDto.addBoard(new BoardResponseDto(board,dto));
 
+        }
 
         return boardListResponseDto;
     }
 
     @Transactional
-    public GlobalResDto updateBoard(Long id, BoardDto boardRequestDto) {
-
+    public GlobalResDto updateBoard(Long id, BoardDto boardRequestDto, Member member) {
 
         Board board = checkBoard(boardRepository, id);
+
+        if (!board.getMember().getUsername().equals(member.getUsername())) {
+            throw new RuntimeException("작성자만 수정이 가능합니다.");
+        }
+
         board.boardUpdate(boardRequestDto);
 
         return new GlobalResDto("게시글 수정이 완료되었습니다.", HttpStatus.OK.value());
 
     }
 
-    // 작성자가 삭제 할수있게끔 해주세요~
-
     @Transactional
-    public GlobalResDto deleteBoard(Long id) {
+    public GlobalResDto deleteBoard(Long id, Member member) {
+
         Board board = checkBoard(boardRepository, id);
+
+        if (!board.getMember().getUsername().equals(member.getUsername())) {
+            throw new RuntimeException("작성자만 삭제가 가능합니다.");
+        }
+
         boardRepository.delete(board);
+
         return new GlobalResDto("게시글 삭제가 완료되었습니다.", HttpStatus.OK.value());
     }
 
@@ -97,13 +114,5 @@ public class BoardService {
                 () -> new RuntimeException("게시글이 존재하지 않습니다.")
         );
     }
-
-    @Transactional(readOnly = true)
-    public Board isPresentPost(Long id) {
-        Optional<Board> optionalBoard = boardRepository.findById(id);
-        return optionalBoard.orElse(null);
-    }
-
-
 }
